Removed some SCT code as we don't need it Fixed some typos New CA icons Use client CS info only if sent Update CS count in shared file process loop

+++ updownclient.h
@@ -1125,7 +1125,7 @@ public:
     {
         return m_incompletepartVer;
     }
-    bool	IsIncPartAvailable(const uint16 iPart) const;
+    bool	IsIncPartAvailable(const UINT iPart) const;
 //<<< WiZaRd::ICS [enkeyDEV]
 //>>> WiZaRd::AntiHideOS [netfinity]
 public:
